Move VariableLister into basic.py
